treebrowser: update for new libiview API
Requires libiview > 0.4.1

Mostly accounts for changed structure and function names. There is a small
addition of functionality, as iv_fetch_episode now takes a file descriptor
instead of a path name, so we must open a file. The devkitpro/libogc doesn't
include an implementation of creat(2), as used by the reference implementation
in iviewiir, so use open(2) with O_CREAT instead. Also, there is no basename(3)
implementation, so the previous commit (39498e1) adds one.